# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Treasure Hunt Szczecin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface GuestLoginBody {
  code: string;
}

export interface AuthResponse {
  id: number;
  name: string;
  role: string;
}

export interface GuestAuthResponse {
  markerId: number;
  code: string;
  markerTitle: string;
  lat: number;
  lng: number;
}

export interface Marker {
  id: number;
  title: string;
  description: string;
  code: string;
  lat: number;
  lng: number;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
  expiresAt: string;
  redemptionCount: number;
}

export interface CreateMarkerBody {
  title: string;
  description: string;
  lat: number;
  lng: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Redemption {
  id: number;
  userId: number;
  markerId: number;
  markerTitle: string;
  redeemedAt: string;
}

export interface CreateRedemptionBody {
  userId: number;
  code: string;
}

export interface Stats {
  totalMarkers: number;
  activeMarkers: number;
  expiredMarkers: number;
  totalRedemptions: number;
}

export type ListRedemptionsParams = {
  userId?: number;
};