# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Treasure Hunt Szczecin API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  role: zod.string(),
});

/**
 * @summary Guest login via treasure code
 */
export const GuestLoginBody = zod.object({
  code: zod.string(),
});

export const GuestLoginResponse = zod.object({
  markerId: zod.number(),
  code: zod.string(),
  markerTitle: zod.string(),
  lat: zod.number(),
  lng: zod.number(),
});

/**
 * @summary List all markers (admin only)
 */
export const ListMarkersResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  code: zod.string(),
  lat: zod.number(),
  lng: zod.number(),
  imageUrl: zod.string().nullable(),
  createdAt: zod.string(),
  expiresAt: zod.string(),
  redemptionCount: zod.number(),
});
export const ListMarkersResponse = zod.array(ListMarkersResponseItem);

/**
 * @summary Create a new treasure marker (admin only)
 */
export const CreateMarkerBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  lat: zod.number(),
  lng: zod.number(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Get a marker by ID
 */
export const GetMarkerParams = zod.object({
  id: zod.coerce.number(),
});

export const GetMarkerResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  code: zod.string(),
  lat: zod.number(),
  lng: zod.number(),
  imageUrl: zod.string().nullable(),
  createdAt: zod.string(),
  expiresAt: zod.string(),
  redemptionCount: zod.number(),
});

/**
 * @summary Delete a marker (admin only)
 */
export const DeleteMarkerParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get a marker by treasure code (for guests)
 */
export const GetMarkerByCodeParams = zod.object({
  code: zod.coerce.string(),
});

export const GetMarkerByCodeResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  code: zod.string(),
  lat: zod.number(),
  lng: zod.number(),
  imageUrl: zod.string().nullable(),
  createdAt: zod.string(),
  expiresAt: zod.string(),
  redemptionCount: zod.number(),
});

/**
 * @summary List redemptions for a user
 */
export const ListRedemptionsQueryParams = zod.object({
  userId: zod.coerce.number().optional(),
});

export const ListRedemptionsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  markerId: zod.number(),
  markerTitle: zod.string(),
  redeemedAt: zod.string(),
});
export const ListRedemptionsResponse = zod.array(ListRedemptionsResponseItem);

/**
 * @summary Redeem a treasure code (regular user)
 */
export const CreateRedemptionBody = zod.object({
  userId: zod.number(),
  code: zod.string(),
});

/**
 * @summary Get overall stats (admin view)
 */
export const GetStatsResponse = zod.object({
  totalMarkers: zod.number(),
  activeMarkers: zod.number(),
  expiredMarkers: zod.number(),
  totalRedemptions: zod.number(),
});